requirements.txt
beautifulsoup4==4.12.2
requests==2.31.0
urllib3==2.0.6
PyMuPDF==1.22.3
pytesseract==0.3.10
Pillow==9.5.0
fitz==1.22.5
numpy==1.26.0
scikit-learn==1.3.1
matplotlib==3.8.0
